# spring-rest/OpenApi/openApi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
tags:
  - name: get All
    description: Get all users
  - name: POST for user
    description: Saves information about new User
  - name: My home controller
    description: This controller is used for user's CRUD operations.
paths:
  /api/user:
    get:
      tags:
        - My home controller
        - get All
      operationId: getAll
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    put:
      tags:
        - My home controller
      operationId: update
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
    post:
      tags:
        - My home controller
        - POST for user
      operationId: addUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: success operation
          headers:
            My header:
              description: just for test
              required: true
              style: simple
          $ref:
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: something broken
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDTO'
  /api/user/{userId}:
    get:
      tags:
        - My home controller
      operationId: get
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags:
        - My home controller
      operationId: delete
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
components:
  schemas:
    User:
      required:
        - password
        - username
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        password:
          type: string
        sex:
          type: string
          enum:
            - MAN
            - WOMAN
        date:
          type: string
          format: date-time
    ErrorDTO:
      type: object
      properties:
        message:
          type: string